requirements.txt
cog>=0.8.0
torch==2.1.2
transformers==4.37.2
diffusers==0.25.0
accelerate==0.26.1
safetensors==0.4.2
pillow==10.1.0
xformers==0.0.23.post1
omegaconf==2.3.0
opencv-python-headless==4.8.1.78